moved class definitions to module:x
