--- CollectionOfInfoAssetIVM.py
@@ -1,5 +1,0 @@
-class CollectionOfAssetIVM:
-    _count = 0
-    def __init__(self):
-        _count += 1
-
