--- ScopeContentVM.py
@@ -1,11 +1,0 @@
-
-
-class ScopeContentViewModel:
-    _count = 0
-    def __init__(self, json):
-    _count +=1
-    self.PlaceNames = [] # Collection of strings
-    self.Description: str = ""
-    self.Ephemera: str = ""
-    self.Schema: str = ""
-
